Integrate patch and extension into core
Patch: static_pages_controller.rb
Extension: nav_bar_item_policy.rb
