+++ back/app/policies/nav_bar_item_policy.rb
@@ -26,6 +26,32 @@ class NavBarItemPolicy < ApplicationPolicy
   def removed_default_items?
     user&.active? && user&.admin?
   end
-end
 
-NavBarItemPolicy.include CustomizableNavbar::Extensions::NavBarItemPolicy
+  def create?
+    user&.active? && user&.admin?
+  end
+
+  def update?
+    user&.active? && user&.admin?
+  end
+
+  def reorder?
+    update?
+  end
+
+  def destroy?
+    update?
+  end
+
+  def permitted_attributes_for_create
+    [:code, :static_page_id, { title_multiloc: CL2_SUPPORTED_LOCALES }]
+  end
+
+  def permitted_attributes_for_update
+    [title_multiloc: CL2_SUPPORTED_LOCALES]
+  end
+
+  def permitted_attributes_for_reorder
+    %i[ordering]
+  end
+end
